agregado imagenes para niveles, reorganizacion de videos
