+++ src/App.jsx
@@ -3,7 +3,7 @@ import { LecSuma } from "./pages/LecSuma";
 import { LecResta } from "./pages/LecResta";
 import { LecMultiplicacion } from "./pages/LecMultiplicacion";
 import { LecDivision } from "./pages/LecDivision";
-import {NivelSemillaSumaV1} from "./videos/NivelSemillaSumaV1";
+import {NivelSemillaSumaV1} from "./videos/VIDEOSUMA/NivelSemillaSumaV1";
 import stylesApp from "./App.module.css";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import suma from "./images/EducateLOGO.png";
@@ -11,62 +11,62 @@ import { Inicio } from "./components/Inicio";
 import { LecSNB } from "./pages/LecSNB";
 import { LecSNP } from "./pages/LecSNP";
 import { LecSNA } from "./pages/LecSNA";
-import {NivelSSV2} from "./videos/NivelSSV2"
-import {NivelSSV3} from "./videos/NivelSSV3"
-import { NBSV1 } from "./videos/NBSV1";
-import { NBSV3 } from "./videos/NBSV3";
-import { NBSV2 } from "./videos/NBSV2";
-import { NPSV1 } from "./videos/NPSV1";
-import { NPSV2 } from "./videos/NPSV2";
-import { NPSV3 } from "./videos/NPSV3";
-import { NASV1 } from "./videos/NASV1";
-import { NASV2 } from "./videos/NASV2";
-import { NASV3 } from "./videos/NASV3";
+import {NivelSSV2} from "./videos/VIDEOSUMA/NivelSSV2"
+import {NivelSSV3} from "./videos/VIDEOSUMA/NivelSSV3"
+import { NBSV1 } from "./videos/VIDEOSUMA/NBSV1";
+import { NBSV3 } from "./videos/VIDEOSUMA/NBSV3";
+import { NBSV2 } from "./videos/VIDEOSUMA/NBSV2";
+import { NPSV1 } from "./videos/VIDEOSUMA/NPSV1";
+import { NPSV2 } from "./videos/VIDEOSUMA/NPSV2";
+import { NPSV3 } from "./videos/VIDEOSUMA/NPSV3";
+import { NASV1 } from "./videos/VIDEOSUMA/NASV1";
+import { NASV2 } from "./videos/VIDEOSUMA/NASV2";
+import { NASV3 } from "./videos/VIDEOSUMA/NASV3";
 import { LecRNB } from "./pages/LecRNB";
 import { LecRNP } from "./pages/LecRNP";
 import { LecRNA } from "./pages/LecRNA";
-import { NSRV1 } from "./videos/NSRV1";
-import { NSRV2 } from "./videos/NSRV2";
-import { NSRV3 } from "./videos/NSRV3";
-import { NBRV1 } from "./videos/NBRV1";
-import { NBRV2 } from "./videos/NBRV2";
-import { NBRV3 } from "./videos/NBRV3";
-import { NPRV1 } from "./videos/NPRV1";
-import { NPRV2 } from "./videos/NPRV2";
-import { NPRV3 } from "./videos/NPRV3";
-import { NARV1 } from "./videos/NARV1";
-import { NARV2 } from "./videos/NARV2";
-import { NARV3 } from "./videos/NARV3";
+import { NSRV1 } from "./videos/VIDEORESTA/NSRV1";
+import { NSRV2 } from "./videos/VIDEORESTA/NSRV2";
+import { NSRV3 } from "./videos/VIDEORESTA/NSRV3";
+import { NBRV1 } from "./videos/VIDEORESTA/NBRV1";
+import { NBRV2 } from "./videos/VIDEORESTA/NBRV2";
+import { NBRV3 } from "./videos/VIDEORESTA/NBRV3";
+import { NPRV1 } from "./videos/VIDEORESTA/NPRV1";
+import { NPRV2 } from "./videos/VIDEORESTA/NPRV2";
+import { NPRV3 } from "./videos/VIDEORESTA/NPRV3";
+import { NARV1 } from "./videos/VIDEORESTA/NARV1";
+import { NARV2 } from "./videos/VIDEORESTA/NARV2";
+import { NARV3 } from "./videos/VIDEORESTA/NARV3";
 import { LecMNB } from "./pages/LecMNB";
 import { LecMNP } from "./pages/LecMNP";
 import { LecMNA } from "./pages/LecMNA";
-import { NSMV1 } from "./videos/NSMV1";
-import { NSMV2 } from "./videos/NSMV2";
-import { NSMV3 } from "./videos/NSMV3";
-import { NBMV1 } from "./videos/NBMV1";
-import { NBMV2 } from "./videos/NBMV2";
-import { NBMV3 } from "./videos/NBMV3";
-import { NPMV1 } from "./videos/NPMV1";
-import { NPMV2 } from "./videos/NPMV2";
-import { NPMV3 } from "./videos/NPMV3";
-import { NAMV1 } from "./videos/NAMV1";
-import { NAMV2 } from "./videos/NAMV2";
-import { NAMV3 } from "./videos/NAMV3";
+import { NSMV1 } from "./videos/VIDEOMULT/NSMV1";
+import { NSMV2 } from "./videos/VIDEOMULT/NSMV2";
+import { NSMV3 } from "./videos/VIDEOMULT/NSMV3";
+import { NBMV1 } from "./videos/VIDEOMULT/NBMV1";
+import { NBMV2 } from "./videos/VIDEOMULT/NBMV2";
+import { NBMV3 } from "./videos/VIDEOMULT/NBMV3";
+import { NPMV1 } from "./videos/VIDEOMULT/NPMV1";
+import { NPMV2 } from "./videos/VIDEOMULT/NPMV2";
+import { NPMV3 } from "./videos/VIDEOMULT/NPMV3";
+import { NAMV1 } from "./videos/VIDEOMULT/NAMV1";
+import { NAMV2 } from "./videos/VIDEOMULT/NAMV2";
+import { NAMV3 } from "./videos/VIDEOMULT/NAMV3";
 import { LecDNB } from "./pages/LecDNB";
 import { LecDNP } from "./pages/LecDNP";
 import { LecDNA } from "./pages/LecDNA";
-import { NSDV1 } from "./videos/NSDV1";
-import { NSDV2 } from "./videos/NSDV2";
-import { NSDV3 } from "./videos/NSDV3";
-import { NBDV1 } from "./videos/NBDV1";
-import { NBDV2 } from "./videos/NBDV2";
-import { NBDV3 } from "./videos/NBDV3";
-import { NPDV1 } from "./videos/NPDV1";
-import { NPDV2 } from "./videos/NPDV2";
-import { NPDV3 } from "./videos/NPDV3";
-import { NADV1 } from "./videos/NADV1";
-import { NADV2 } from "./videos/NADV2";
-import { NADV3 } from "./videos/NADV3";
+import { NSDV1 } from "./videos/VIDEODIVIS/NSDV1";
+import { NSDV2 } from "./videos/VIDEODIVIS/NSDV2";
+import { NSDV3 } from "./videos/VIDEODIVIS/NSDV3";
+import { NBDV1 } from "./videos/VIDEODIVIS/NBDV1";
+import { NBDV2 } from "./videos/VIDEODIVIS/NBDV2";
+import { NBDV3 } from "./videos/VIDEODIVIS/NBDV3";
+import { NPDV1 } from "./videos/VIDEODIVIS/NPDV1";
+import { NPDV2 } from "./videos/VIDEODIVIS/NPDV2";
+import { NPDV3 } from "./videos/VIDEODIVIS/NPDV3";
+import { NADV1 } from "./videos/VIDEODIVIS/NADV1";
+import { NADV2 } from "./videos/VIDEODIVIS/NADV2";
+import { NADV3 } from "./videos/VIDEODIVIS/NADV3";
 
 export function App() {
   return (
